refactor(cfg): кэш графов переехал в пакет cfg
Кэш стоит рядом с построением графа, а не в отдельном пакете: `ControlFlowGraphIndex`
вернулся в `cfg`. В карте слоёв `Cfg` перестал быть листом — ему разрешены `Context`
(документ, чей разбор кэшируется), `Index` (база сброса по жизненному циклу) и
`Infrastructure` (workspace-скоуп). В `index` осталась только сама база.

`CfgBuildOptions.buildGraph` снова не публичный: единственный потребитель — кэш из
того же пакета.
