+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/cfg/CLAUDE.md
@@ -33,10 +33,12 @@
 - **`CfgBuildOptions`** — неизменяемый набор настроек (`loopIterations`,
   `preprocessorConditions`, `adjacentDeadCode`) плюс `buildGraph(codeBlock)`. Все три
   настройки передаются построителю явно, поэтому набор задаёт структуру графа целиком.
-- Готовый граф берётся **не построением на месте**, а из
-  `index/ControlFlowGraphIndex.graphOf(documentContext, codeBlock, options)`: он строит
-  граф лениво, делит его между потребителями и сбрасывает по жизненному циклу документа.
-  Выдаваемый экземпляр общий — **только для чтения**.
+- **`ControlFlowGraphIndex`** (workspace-scoped) — источник готовых графов:
+  `graphOf(documentContext, codeBlock, options)` строит граф лениво, делит его между
+  потребителями и сбрасывает по жизненному циклу документа (база — `index/`
+  `AbstractDocumentLifecycleClearableIndex`). Ключ — пара «блок кода + настройки»;
+  блок сравнивается по тождественности. Выдаваемый экземпляр общий — **только для чтения**.
+  Потребители берут граф отсюда, а не строят на месте.
 
 ## Потребители
 
@@ -49,6 +51,6 @@
   `CfgBuildingParseTreeVisitor`; при добавлении типа ребра расширь `CfgEdgeType` и оба обходчика.
 - Новая настройка построения = поле в `CfgBuildOptions` (она же — часть ключа кэша), а не
   очередной сеттер, выставляемый потребителем самостоятельно.
-- Пакет — лист в карте слоёв (`ArchitectureTest`): он ничего не знает про документы, контекст
-  и Spring. Всё, что требует этих знаний, живёт в `index/`.
+- Знание о документах живёт только в `ControlFlowGraphIndex`; само построение графа
+  (`CfgBuildingParseTreeVisitor`, вершины, рёбра) про `DocumentContext` и Spring не знает.
 - Граф — на JGraphT; используй его API, не изобретай обход с нуля.
