+++ src/test/java/com/github/_1c_syntax/bsl/languageserver/cfg/ControlFlowGraphIndexTest.java
@@ -19,9 +19,8 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with BSL Language Server.
  */
-package com.github._1c_syntax.bsl.languageserver.index;
+package com.github._1c_syntax.bsl.languageserver.cfg;
 
-import com.github._1c_syntax.bsl.languageserver.cfg.CfgBuildOptions;
 import com.github._1c_syntax.bsl.languageserver.context.DocumentContext;
 import com.github._1c_syntax.bsl.languageserver.context.events.DocumentContextContentChangedEvent;
 import com.github._1c_syntax.bsl.parser.BSLParser;
